backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
sqlmodel==0.0.24
SQLAlchemy==2.0.43
pillow==11.3.0
